refactor: moving product validation outside of the main form object
